requirements.txt
langgraph>=0.2.27
openai>=1.51.0
requests>=2.32.3
python-dotenv>=1.0.1
pyyaml>=6.0.2
datasets>=3.0.1
tenacity>=8.5.0
pydantic>=2.9.2
rich>=13.9.2
numpy>=1.26.4
transformers>=4.44.2
accelerate>=0.34.2
bitsandbytes>=0.43.3
torch>=2.3.0
gradio>=4.44.0

